Rename namespaces Egui.EguiExtras -> Egui.Extras, Egui.EguiPlot -> Egui.Plot
Both namespaces were derived directly from their crate names
("egui_extras", "egui_plot") PascalCased, picking up the redundant "Egui"
prefix every other autogenerated namespace already avoids (compare
Egui.Widgets, not Egui.EguiWidgets). Renamed via the same
NAMESPACE_RENAMES mechanism already used for the other namespace
collapses, plus the matching directory/using-statement updates for the
hand-written files that live in these namespaces (Plot.cs/PlotUi.cs/
PlotResponse.cs, the Strip/Table family).
