+++ examples/SilkOpenGl/Program.cs
@@ -6,9 +6,9 @@ using System.Collections.Immutable;
 using Egui;
 using Egui.Containers;
 using Egui.Epaint;
-using Egui.EguiExtras;
-using Egui.EguiExtras.SyntaxHighlighting;
-using Egui.EguiPlot;
+using Egui.Extras;
+using Egui.Extras.SyntaxHighlighting;
+using Egui.Plot;
 using Egui.Silk.NET;
 using Egui.Viewport;
 using Egui.Widgets;
